Allow input image to influence noise generation
In previous commits the op did not allow input image data to influence
the output values. Now the algorithm allows, given a nonzero input image
(in which all of the pixels do not have to be the same value), the input
image to influence the output noise, i.e. if it is desired to add
uniform noise with range [-4, 4] upon an imported image, the algorithm
will now add a value x such that -4 <= x <= 4 to each pixel in the
image, whatever value that may be, as long as it does not extend beyond
the range of the type.