Fix Img creation when called as function
Previously, when DefaultFloodFill was called as a function its
initialize() method used CreateImgFromInterval, which could be
problematic for the Op if the input image was not also DoubleType
since that particular create Op returned a DoubleType Img. The Create Op
was replaced with CreateImgFromRAI, which returns an Img that is of the
same Type, avoiding this error.